refactor route courses server
