+++ server/routes/academy.js
@@ -6,521 +6,6 @@ const { body, validationResult, check } = require('express-validator');
 const uuidv4 = require('uuid/v4');
 const Status = { Open: 'Open', InProgress: 'InProgress', Completed: 'Completed' };
 
-// Edit course info
-router.put(
-  '/course',
-  checkJWT,
-  [
-    body('courseId')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape(),
-    body('courseCode')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape(),
-    body('courseName')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape(),
-    body('shortDescription')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape(),
-    body('description')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape()
-  ],
-  async (req, res) => {
-    if (req.decoded.user.role !== USER_ROLES.ADMIN_ACADEMY) {
-      return res.status(403).json({
-        success: false,
-        msg: 'Permission Denied'
-      });
-    } else {
-      const errors = validationResult(req);
-
-      if (!errors.isEmpty()) {
-        return res.status(422).json({ errors: errors.array() });
-      }
-
-      const user = req.decoded.user;
-      const networkObj = await network.connectToNetwork(user);
-
-      const { courseId, courseCode, courseName, shortDescription, description } = req.body;
-
-      let course = {
-        courseId: courseId,
-        courseCode: courseCode,
-        courseName: courseName,
-        shortDescription: shortDescription,
-        description: description
-      };
-
-      const response = await network.updateCourseInfo(networkObj, course);
-
-      if (!response.success) {
-        return res.status(500).json({
-          success: false,
-          msg: response.msg
-        });
-      }
-
-      const listNewCourse = await network.query(networkObj, 'GetAllCourses');
-
-      return res.json({
-        success: true,
-        courses: JSON.parse(listNewCourse.msg)
-      });
-    }
-  }
-);
-
-// Subject of course
-router.post(
-  '/addSubjectToCourse',
-  checkJWT,
-  [
-    body('courseId')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape(),
-    body('subjectId')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape()
-  ],
-  async (req, res) => {
-    if (req.decoded.user.role !== USER_ROLES.ADMIN_ACADEMY) {
-      return res.status(403).json({
-        success: false,
-        msg: 'Permission Denied'
-      });
-    }
-    let errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      return res.status(422).json({ success: false, errors: errors.array() });
-    }
-
-    let admin = req.decoded.user;
-    let networkObj = await network.connectToNetwork(admin);
-    if (!networkObj) {
-      return res.status(500).json({
-        msg: 'Failed connect to blockchain!'
-      });
-    }
-
-    let courseId = req.body.courseId;
-    let subjectId = req.body.subjectId;
-
-    let course = await network.query(networkObj, 'GetCourse', courseId);
-    if (!course.success) {
-      return res.status(404).json({
-        msg: 'Can not query chaincode!'
-      });
-    }
-
-    course = JSON.parse(course.msg);
-
-    if (course.Status === 'Closed') {
-      return res.status(400).json({
-        msg: 'This course was closed!'
-      });
-    }
-
-    if (course.Subjects && course.Subjects.includes(subjectId)) {
-      return res.status(409).json({
-        msg: 'This subject already presents in course!'
-      });
-    }
-
-    networkObj = await network.connectToNetwork(admin);
-    let response = await network.addSubjectToCourse(networkObj, courseId, subjectId);
-
-    if (!response.success) {
-      return res.status(500).json({
-        msg: 'Can not invoke chaincode!'
-      });
-    }
-
-    return res.status(200).json({
-      msg: 'Add Sucessfully!'
-    });
-  }
-);
-router.post(
-  '/removeSubjectFromCourse',
-  checkJWT,
-  [
-    body('courseId')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape(),
-    body('subjectId')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape()
-  ],
-  async (req, res) => {
-    if (req.decoded.user.role !== USER_ROLES.ADMIN_ACADEMY) {
-      return res.status(403).json({
-        msg: 'Permission Denied'
-      });
-    }
-    let errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      return res.status(422).json({ errors: errors.array() });
-    }
-
-    let user = req.decoded.user;
-    let networkObj = await network.connectToNetwork(user);
-    if (!networkObj) {
-      return res.status(500).json({
-        msg: 'Failed connect to blockchain!'
-      });
-    }
-
-    let courseId = req.body.courseId;
-    let subjectId = req.body.subjectId;
-
-    let query = await network.query(networkObj, 'GetCourse', courseId);
-
-    if (!query.success) {
-      return res.status(404).json({
-        success: false,
-        msg: 'Can not query chaincode!'
-      });
-    }
-
-    let course = JSON.parse(query.msg);
-
-    if (!course.Subjects || course.Subjects.indexOf(subjectId) === -1) {
-      return res.status(404).json({
-        msg: 'This subject does not present in course!'
-      });
-    }
-    networkObj = await network.connectToNetwork(user);
-
-    let response = await network.removeSubjectFromCourse(networkObj, courseId, subjectId);
-    if (!response.success) {
-      return res.status(500).json({
-        msg: 'Can not invoke chaincode!'
-      });
-    }
-
-    return res.json({
-      msg: 'This subject has been removed from course!'
-    });
-  }
-);
-
-router.get(
-  '/subjectNoCourse/:courseId',
-  checkJWT,
-  check('courseId')
-    .trim()
-    .escape(),
-  async (req, res) => {
-    const courseId = req.params.courseId;
-
-    if (req.decoded.user.role !== USER_ROLES.ADMIN_ACADEMY) {
-      return res.status(403).json({
-        msg: 'Permission Denied'
-      });
-    }
-    let errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ errors: errors.array() });
-    }
-
-    const user = req.decoded.user;
-    const networkObj = await network.connectToNetwork(user);
-    if (!networkObj) {
-      return res.status(500).json({
-        msg: 'Failed connect to blockchain'
-      });
-    }
-
-    const subjectsNotInCourse = await network.query(networkObj, 'GetSubjectsNotInCourse', courseId);
-    if (!subjectsNotInCourse.success) {
-      return res.status(404).send({
-        msg: 'Can not query chaincode!'
-      });
-    }
-
-    return res.status(200).json({
-      subjects: JSON.parse(subjectsNotInCourse.msg)
-    });
-  }
-);
-
-// Course Manager
-
-router.post(
-  '/course',
-  checkJWT,
-  [
-    body('courseCode')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape(),
-    body('courseName')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape(),
-    body('shortDescription')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape(),
-    body('description')
-      .not()
-      .isEmpty()
-      .trim()
-      .escape()
-  ],
-  async (req, res) => {
-    if (req.decoded.user.role !== USER_ROLES.ADMIN_ACADEMY) {
-      return res.status(403).json({
-        success: false,
-        msg: 'Permission Denied'
-      });
-    } else {
-      const errors = validationResult(req);
-
-      if (!errors.isEmpty()) {
-        return res.status(422).json({ errors: errors.array() });
-      }
-
-      const user = req.decoded.user;
-      const networkObj = await network.connectToNetwork(user);
-
-      const { courseCode, courseName, shortDescription, description } = req.body;
-
-      let course = {
-        courseId: uuidv4(),
-        courseCode: courseCode,
-        courseName: courseName,
-        shortDescription: shortDescription,
-        description: description
-      };
-
-      const response = await network.createCourse(networkObj, course);
-
-      if (!response.success) {
-        return res.status(500).json({
-          success: false,
-          msg: response.msg
-        });
-      }
-
-      const listNewCourse = await network.query(networkObj, 'GetAllCourses');
-
-      return res.json({
-        success: true,
-        courses: JSON.parse(listNewCourse.msg)
-      });
-    }
-  }
-);
-
-router.post(
-  '/closeCourse',
-  checkJWT,
-  body('courseId')
-    .not()
-    .isEmpty()
-    .trim()
-    .escape(),
-  async (req, res) => {
-    if (req.decoded.user.role !== USER_ROLES.ADMIN_ACADEMY) {
-      return res.status(403).json({
-        success: false,
-        msg: 'Permission Denied'
-      });
-    }
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      return res.status(422).json({
-        success: false,
-        errors: errors.array()
-      });
-    }
-
-    let user = req.decoded.user;
-    let networkObj = await network.connectToNetwork(user);
-    if (!networkObj) {
-      return res.status(500).json({
-        success: false,
-        msg: 'Failed connect to blockchain!'
-      });
-    }
-
-    let { courseId } = req.body;
-    let query = await network.query(networkObj, 'GetCourse', courseId);
-    if (!query.success) {
-      return res.status(500).json({
-        success: false,
-        msg: 'Can not query chaincode!'
-      });
-    }
-
-    courseInfo = JSON.parse(query.msg);
-    if (courseInfo.Status === 'Closed') {
-      return res.status(500).json({
-        success: false,
-        msg: 'This course was closed!'
-      });
-    }
-
-    networkObj = await network.connectToNetwork(user);
-
-    let response = await network.closeCourse(networkObj, courseId);
-
-    if (!response.success) {
-      return res.status(500).json({
-        success: false,
-        msg: 'Can not invoke chaincode!'
-      });
-    }
-
-    return res.status(200).json({
-      success: true,
-      msg: 'Close Successfully!'
-    });
-  }
-);
-
-router.post(
-  '/openCourse',
-  checkJWT,
-  body('courseId')
-    .not()
-    .isEmpty()
-    .trim()
-    .escape(),
-  async (req, res) => {
-    if (req.decoded.user.role !== USER_ROLES.ADMIN_ACADEMY) {
-      return res.status(403).json({
-        success: false,
-        msg: 'Permission Denied'
-      });
-    }
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      return res.status(422).json({
-        success: false,
-        errors: errors.array()
-      });
-    }
-
-    let user = req.decoded.user;
-    let networkObj = await network.connectToNetwork(user);
-    if (!networkObj) {
-      return res.status(500).json({
-        success: false,
-        msg: 'Failed connect to blockchain!'
-      });
-    }
-
-    let { courseId } = req.body;
-    let query = await network.query(networkObj, 'GetCourse', courseId);
-    if (!query.success) {
-      return res.status(500).json({
-        success: false,
-        msg: 'Can not query chaincode!'
-      });
-    }
-
-    courseInfo = JSON.parse(query.msg);
-    if (courseInfo.Status === 'Open') {
-      return res.status(500).json({
-        success: false,
-        msg: 'This course is open!'
-      });
-    }
-
-    networkObj = await network.connectToNetwork(user);
-
-    let response = await network.openCourse(networkObj, courseId);
-
-    if (!response.success) {
-      return res.status(500).json({
-        success: false,
-        msg: 'Can not invoke chaincode!'
-      });
-    }
-
-    return res.status(200).json({
-      success: true,
-      msg: 'Open Successfully!'
-    });
-  }
-);
-
-router.get(
-  '/coursesOfStudent/:username',
-  checkJWT,
-  check('username')
-    .trim()
-    .escape(),
-  async (req, res) => {
-    const user = req.decoded.user;
-
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      return res.status(422).json({ errors: errors.array() });
-    }
-
-    if (user.role !== USER_ROLES.ADMIN_ACADEMY) {
-      return res.status(403).json({
-        success: false,
-        msg: 'Permission Denied'
-      });
-    }
-
-    const networkObj = await network.connectToNetwork(user);
-
-    if (!networkObj) {
-      return res.status(500).json({
-        success: false,
-        msg: 'Connect to blockchain failed'
-      });
-    }
-
-    const response = await network.query(networkObj, 'GetCoursesOfStudent', req.params.username);
-
-    if (!response.success) {
-      return res.status(500).json({
-        success: false,
-        msg: 'Query chaincode failed'
-      });
-    }
-
-    return res.json({
-      success: true,
-      courses: JSON.parse(response.msg)
-    });
-  }
-);
 //get class of student
 router.get(
   '/classesOfStudent/:username',
@@ -726,6 +211,7 @@ router.put(
   async (req, res) => {
     if (req.decoded.user.role !== USER_ROLES.ADMIN_ACADEMY) {
       return res.status(403).json({
+        success: false,
         msg: 'Permission Denied'
       });
     } else {
@@ -765,6 +251,7 @@ router.put(
 
       if (!response.success) {
         return res.status(500).json({
+          success: false,
           msg: response.msg
         });
       }
@@ -772,6 +259,7 @@ router.put(
       const listNewClass = await network.query(networkObj, 'GetClassesOfSubject', subjectId);
 
       return res.json({
+        success: true,
         classes: JSON.parse(listNewClass.msg)
       });
     }
@@ -1288,88 +776,4 @@ router.post(
   }
 );
 
-//get Student by Id
-router.get(
-  '/student/:username',
-  check('username')
-    .trim()
-    .escape(),
-  checkJWT,
-  async (req, res) => {
-    if (req.decoded.user.role !== USER_ROLES.ADMIN_ACADEMY) {
-      return res.status(403).json({
-        success: false,
-        msg: 'Permission Denied'
-      });
-    }
-    const user = req.decoded.user;
-
-    const networkObj = await network.connectToNetwork(user);
-
-    if (!networkObj) {
-      return res.status(500).json({
-        success: false,
-        msg: 'Failed connect to blockchain!'
-      });
-    }
-
-    const response = await network.query(networkObj, 'GetStudent', req.params.username);
-
-    if (!response.success) {
-      return res.status(500).send({
-        success: false,
-        msg: response.msg.toString()
-      });
-    }
-
-    let student = JSON.parse(response.msg);
-
-    return res.json({
-      success: true,
-      student: student
-    });
-  }
-);
-// get teacher by username
-router.get(
-  '/teacher/:username',
-  check('username')
-    .trim()
-    .escape(),
-  checkJWT,
-  async (req, res) => {
-    if (req.decoded.user.role !== USER_ROLES.ADMIN_ACADEMY) {
-      return res.status(403).json({
-        success: false,
-        msg: 'Permission Denied'
-      });
-    }
-    const user = req.decoded.user;
-
-    const networkObj = await network.connectToNetwork(user);
-
-    if (!networkObj) {
-      return res.status(500).json({
-        success: false,
-        msg: 'Failed connect to blockchain!'
-      });
-    }
-
-    const response = await network.query(networkObj, 'GetTeacher', req.params.username);
-
-    if (!response.success) {
-      return res.status(500).send({
-        success: false,
-        msg: response.msg.toString()
-      });
-    }
-
-    let teacher = JSON.parse(response.msg) ? JSON.parse(response.msg) : {};
-
-    return res.json({
-      success: true,
-      teacher: teacher
-    });
-  }
-);
 module.exports = router;
