+++ server/test/academy.test.js
@@ -10,647 +10,6 @@ const User = require('../models/User');
 
 require('dotenv').config();
 
-describe('#PUT /academy/course', () => {
-  let connect;
-  let query;
-  let updateCourseInfo;
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    query = sinon.stub(network, 'query');
-    updateCourseInfo = sinon.stub(network, 'updateCourseInfo');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    query.restore();
-    updateCourseInfo.restore();
-  });
-
-  it('permission denied when access routes with student', (done) => {
-    request(app)
-      .put('/academy/course')
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('permission denied when access routes with teacher', (done) => {
-    request(app)
-      .put('/academy/course')
-      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('success edit course', (done) => {
-    let data = JSON.stringify(
-      {
-        CourseID: '123-456-789',
-        CourseName: 'Blockchain101',
-        CourseCode: 'BC101',
-        Description: 'Blockchain'
-      },
-      {
-        CourseID: '123-456-783',
-        CourseName: 'Blockchain102',
-        CourseCode: 'BC102',
-        Description: 'Blockchain'
-      }
-    );
-
-    updateCourseInfo.returns({
-      success: true,
-      msg: 'Edit success!'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .put('/academy/course')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '123-456-789',
-        courseCode: 'BC101',
-        courseName: 'Blockchain 101',
-        shortDescription: 'Blockchain',
-        description: 'Blockchain for beginer'
-      })
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
-        done();
-      });
-  });
-
-  it('do not success edit course because req.body invalid', (done) => {
-    request(app)
-      .put('/academy/course')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '',
-        courseCode: 'BC101',
-        courseName: 'Blockchain 101',
-        description: 'Blockchain'
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        done();
-      });
-  });
-
-  it('edit course info fail when call editCouseInfo function', (done) => {
-    updateCourseInfo.returns({
-      success: false,
-      msg: 'Error chaincode'
-    });
-
-    request(app)
-      .put('/academy/course')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '123-456-789',
-        courseCode: 'BC101',
-        courseName: 'Blockchain 101',
-        shortDescription: 'Blockchain',
-        description: 'Blockchain for beginer'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-});
-
-describe('#POST /academy/course', () => {
-  let connect;
-  let query;
-  let createCourse;
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    query = sinon.stub(network, 'query');
-    createCourse = sinon.stub(network, 'createCourse');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    query.restore();
-    createCourse.restore();
-  });
-
-  it('permission denied when access routes with student', (done) => {
-    request(app)
-      .post('/academy/course')
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('permission denied when access routes with teacher', (done) => {
-    request(app)
-      .post('/academy/course')
-      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('success create courses', (done) => {
-    let data = JSON.stringify(
-      {
-        CourseID: '123-456-789',
-        CourseName: 'Blockchain101',
-        CourseCode: 'BC101',
-        Description: 'Blockchain'
-      },
-      {
-        CourseID: '123-456-783',
-        CourseName: 'Blockchain102',
-        CourseCode: 'BC102',
-        Description: 'Blockchain'
-      }
-    );
-
-    createCourse.returns({
-      success: true,
-      msg: 'Create success!'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .post('/academy/course')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseCode: 'BC101',
-        courseName: 'Blockchain 101',
-        shortDescription: 'Blockchain',
-        description: 'Full course blockchain'
-      })
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
-        done();
-      });
-  });
-
-  it('do not success create course because req.body invalid', (done) => {
-    request(app)
-      .post('/academy/course')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseCode: '',
-        courseName: 'Blockchain 101',
-        description: 'Blockchain'
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        done();
-      });
-  });
-
-  it('create course fail when call createCourse function', (done) => {
-    createCourse.returns({
-      success: false,
-      msg: 'Error'
-    });
-
-    request(app)
-      .post('/academy/course')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseCode: 'BC101',
-        courseName: 'Blockchain 101',
-        shortDescription: 'Blockchain',
-        description: 'Blockchain 101 for beginer'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-});
-
-describe('#POST /academy/closeCourse', () => {
-  let connect;
-  let query;
-  let closeCourse;
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    query = sinon.stub(network, 'query');
-    closeCourse = sinon.stub(network, 'closeCourse');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    query.restore();
-    closeCourse.restore();
-  });
-
-  it('permission denied when access routes with student', (done) => {
-    request(app)
-      .post('/academy/closeCourse')
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('permission denied when access routes with teacher', (done) => {
-    request(app)
-      .post('/academy/closeCourse')
-      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('invalid body', (done) => {
-    request(app)
-      .post('/academy/closeCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: ''
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('Failed connect to blockchain!', (done) => {
-    connect.returns(null);
-    request(app)
-      .post('/academy/closeCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Failed connect to blockchain!');
-        done();
-      });
-  });
-
-  it('Can not query chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    query.returns({
-      success: false
-    });
-
-    request(app)
-      .post('/academy/closeCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Can not query chaincode!');
-        done();
-      });
-  });
-
-  it('This course was closed!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify({
-      CourseID: '123456',
-      Status: 'Closed'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .post('/academy/closeCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('This course was closed!');
-        done();
-      });
-  });
-
-  it('Can not invoke chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify({
-      CourseID: '123456',
-      Status: 'Open'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    closeCourse.returns({
-      success: false
-    });
-
-    request(app)
-      .post('/academy/closeCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Can not invoke chaincode!');
-        done();
-      });
-  });
-
-  it('Close Successfully!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify({
-      CourseID: '123456',
-      Status: 'Open'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    closeCourse.returns({
-      success: true
-    });
-
-    request(app)
-      .post('/academy/closeCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
-        expect(res.body.msg).equal('Close Successfully!');
-        done();
-      });
-  });
-});
-
-describe('#POST /academy/openCourse', () => {
-  let connect;
-  let query;
-  let openCourse;
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    query = sinon.stub(network, 'query');
-    openCourse = sinon.stub(network, 'openCourse');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    query.restore();
-    openCourse.restore();
-  });
-
-  it('permission denied when access routes with student', (done) => {
-    request(app)
-      .post('/academy/openCourse')
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('permission denied when access routes with teacher', (done) => {
-    request(app)
-      .post('/academy/openCourse')
-      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('invalid body', (done) => {
-    request(app)
-      .post('/academy/openCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: ''
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('Failed connect to blockchain!', (done) => {
-    connect.returns(null);
-    request(app)
-      .post('/academy/openCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Failed connect to blockchain!');
-        done();
-      });
-  });
-
-  it('Can not query chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    query.returns({
-      success: false
-    });
-
-    request(app)
-      .post('/academy/openCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Can not query chaincode!');
-        done();
-      });
-  });
-
-  it('This course is open!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify({
-      CourseID: '123456',
-      Status: 'Open'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .post('/academy/openCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('This course is open!');
-        done();
-      });
-  });
-
-  it('Can not invoke chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify({
-      CourseID: '123456',
-      Status: 'Closed'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    openCourse.returns({
-      success: false
-    });
-
-    request(app)
-      .post('/academy/openCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Can not invoke chaincode!');
-        done();
-      });
-  });
-
-  it('Close Successfully!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify({
-      CourseID: '123456',
-      Status: 'Closed'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    openCourse.returns({
-      success: true
-    });
-
-    request(app)
-      .post('/academy/openCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
-        expect(res.body.msg).equal('Open Successfully!');
-        done();
-      });
-  });
-});
-
 describe('#POST /academy/:subjectId/class', () => {
   let connect;
   let query;
@@ -840,6 +199,7 @@ describe('#PUT /class', () => {
       .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
       .then((res) => {
         expect(res.status).equal(403);
+        expect(res.body.success).equal(false);
         done();
       });
   });
@@ -850,6 +210,7 @@ describe('#PUT /class', () => {
       .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
       .then((res) => {
         expect(res.status).equal(403);
+        expect(res.body.success).equal(false);
         done();
       });
   });
@@ -893,6 +254,7 @@ describe('#PUT /class', () => {
       })
       .then((res) => {
         expect(res.status).equal(200);
+        expect(res.body.success).equal(true);
         done();
       });
   });
@@ -937,6 +299,7 @@ describe('#PUT /class', () => {
       })
       .then((res) => {
         expect(res.status).equal(500);
+        expect(res.body.success).equal(false);
         done();
       });
   });
@@ -1162,553 +525,6 @@ describe('#POST /academy/subject', () => {
       .then((res) => {
         expect(res.status).equal(200);
         expect(res.body.success).equal(true);
-        done();
-      });
-  });
-});
-
-describe('#POST /academy/addSubjectToCourse', () => {
-  let connect;
-  let query;
-  let addSubjectToCourse;
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    query = sinon.stub(network, 'query');
-    addSubjectToCourse = sinon.stub(network, 'addSubjectToCourse');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    query.restore();
-    addSubjectToCourse.restore();
-  });
-
-  it('permission denied when access routes with student', (done) => {
-    request(app)
-      .post('/academy/addSubjectToCourse')
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('permission denied when access routes with teacher', (done) => {
-    request(app)
-      .post('/academy/addSubjectToCourse')
-      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('Add subject to course failed because req.body invalid', (done) => {
-    request(app)
-      .post('/academy/addSubjectToCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '',
-        subjectId: ''
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        done();
-      });
-  });
-
-  it('Failed connect to blockchain!', (done) => {
-    connect.returns(null);
-    request(app)
-      .post('/academy/addSubjectToCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: 'course01',
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.msg).equal('Failed connect to blockchain!');
-        done();
-      });
-  });
-
-  it('Can not query chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let course = JSON.stringify({
-      CourseID: 'course01',
-      Status: 'Closed'
-    });
-
-    query.returns({
-      success: false
-    });
-
-    request(app)
-      .post('/academy/addSubjectToCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: 'course01',
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(404);
-        expect(res.body.msg).equal('Can not query chaincode!');
-        done();
-      });
-  });
-
-  it('This course was closed!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let course = JSON.stringify({
-      CourseID: 'course01',
-      Status: 'Closed'
-    });
-
-    query.returns({
-      success: true,
-      msg: course
-    });
-
-    request(app)
-      .post('/academy/addSubjectToCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: 'course01',
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(400);
-        expect(res.body.msg).equal('This course was closed!');
-        done();
-      });
-  });
-
-  it('This subject already presents in course!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let course = JSON.stringify({
-      CourseID: 'course01',
-      Status: 'Open',
-      Subjects: ['subject01']
-    });
-
-    query.returns({
-      success: true,
-      msg: course
-    });
-
-    request(app)
-      .post('/academy/addSubjectToCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: 'course01',
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(409);
-        expect(res.body.msg).equal('This subject already presents in course!');
-        done();
-      });
-  });
-
-  it('Can not invoke chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let course = JSON.stringify({
-      CourseID: 'course01',
-      Status: 'Open'
-    });
-
-    query.returns({
-      success: true,
-      msg: course
-    });
-
-    addSubjectToCourse.returns({
-      success: false
-    });
-
-    request(app)
-      .post('/academy/addSubjectToCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: 'course01',
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.msg).equal('Can not invoke chaincode!');
-        done();
-      });
-  });
-
-  it('Add Sucessfully!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let course = JSON.stringify({
-      CourseID: 'course01',
-      Status: 'Open'
-    });
-
-    query.returns({
-      success: true,
-      msg: course
-    });
-
-    addSubjectToCourse.returns({
-      success: true
-    });
-
-    request(app)
-      .post('/academy/addSubjectToCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: 'course01',
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.msg).equal('Add Sucessfully!');
-        done();
-      });
-  });
-});
-
-describe('#POST /academy/removeSubjectFromCourse', () => {
-  let connect;
-  let query;
-  let removeSubjectFromCourse;
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    query = sinon.stub(network, 'query');
-    removeSubjectFromCourse = sinon.stub(network, 'removeSubjectFromCourse');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    query.restore();
-    removeSubjectFromCourse.restore();
-  });
-
-  it('Permission denied when access routes with student', (done) => {
-    request(app)
-      .post('/academy/removeSubjectFromCourse')
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        done();
-      });
-  });
-
-  it('Permission denied when access routes with teacher', (done) => {
-    request(app)
-      .post('/academy/removeSubjectFromCourse')
-      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        done();
-      });
-  });
-
-  it('Remove subject from course failed because req.body invalid', (done) => {
-    request(app)
-      .post('/academy/removeSubjectFromCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '',
-        subjectId: ''
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        done();
-      });
-  });
-
-  it('Failed connect to blockchain!', (done) => {
-    connect.returns(null);
-    request(app)
-      .post('/academy/removeSubjectFromCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: 'course01',
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.msg).equal('Failed connect to blockchain!');
-        done();
-      });
-  });
-
-  it('Can not query chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    query.returns({
-      success: false
-    });
-
-    request(app)
-      .post('/academy/removeSubjectFromCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: 'course01',
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(404);
-        expect(res.body.msg).equal('Can not query chaincode!');
-        done();
-      });
-  });
-
-  it('This subject does not present in course!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify({
-      CourseID: 'e4b5316d-cb22-4e6f-acfd-6ea0feb5478f',
-      Subjects: []
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .post('/academy/removeSubjectFromCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: 'e4b5316d-cb22-4e6f-acfd-6ea0feb5478f',
-        subjectId: 'baa7df8c-8aa1-4ce0-9a26-6439ab11b8ef'
-      })
-      .then((res) => {
-        expect(res.status).equal(404);
-        expect(res.body.msg).equal('This subject does not present in course!');
-        done();
-      });
-  });
-
-  it('Can not invoke chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify({
-      CourseID: 'e4b5316d-cb22-4e6f-acfd-6ea0feb5478f',
-      Subjects: ['baa7df8c-8aa1-4ce0-9a26-6439ab11b8ef']
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    removeSubjectFromCourse.returns({
-      success: false
-    });
-
-    request(app)
-      .post('/academy/removeSubjectFromCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: 'e4b5316d-cb22-4e6f-acfd-6ea0feb5478f',
-        subjectId: 'baa7df8c-8aa1-4ce0-9a26-6439ab11b8ef'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.msg).equal('Can not invoke chaincode!');
-        done();
-      });
-  });
-
-  it('This subject has been removed from course!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify({
-      CourseID: 'e4b5316d-cb22-4e6f-acfd-6ea0feb5478f',
-      Subjects: ['baa7df8c-8aa1-4ce0-9a26-6439ab11b8ef']
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    removeSubjectFromCourse.returns({
-      success: true,
-      msg: 'This subject has been removed from course!'
-    });
-
-    request(app)
-      .post('/academy/removeSubjectFromCourse')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: 'e4b5316d-cb22-4e6f-acfd-6ea0feb5478f',
-        subjectId: 'baa7df8c-8aa1-4ce0-9a26-6439ab11b8ef'
-      })
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.msg).equal('This subject has been removed from course!');
-        done();
-      });
-  });
-});
-
-describe('#GET /academy/subjectNoCourse/:courseId', () => {
-  let connect;
-  let query;
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    query = sinon.stub(network, 'query');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    query.restore();
-  });
-
-  it('do not get success subjects with role admin student', (done) => {
-    request(app)
-      .get('/academy/subjectNoCourse/:courseId')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        courseId: '1'
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-
-  it('do not get success subjects with role teacher', (done) => {
-    request(app)
-      .get('/academy/subjectNoCourse/:courseId')
-      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
-      .send({
-        courseId: '1'
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-
-  it('do not get success subjects with role student', (done) => {
-    request(app)
-      .get('/academy/subjectNoCourse/:courseId')
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .send({
-        courseId: '1'
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-
-  it('Can not connect to network', (done) => {
-    connect.returns(null);
-    request(app)
-      .get('/academy/subjectNoCourse/:courseId')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        courseId: '1'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-
-        expect(res.body.msg).equal('Failed connect to blockchain');
-        done();
-      });
-  });
-
-  it('get success ', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let course = JSON.stringify({
-      CourseID: '12312124123421',
-      Subjects: []
-    });
-
-    let subjects = JSON.stringify([
-      {
-        SubjectID: '1321321'
-      },
-      { SubjectID: '163654654654' }
-    ]);
-
-    query.onFirstCall().returns({ success: true, msg: course });
-    query.onSecondCall().returns({ success: true, msg: subjects });
-    query.onThirdCall().returns({ success: true, msg: subjects });
-
-    request(app)
-      .get('/academy/subjectNoCourse/12312124123421')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({})
-      .then((res) => {
-        expect(res.status).equal(200);
         done();
       });
   });
@@ -3211,124 +2027,6 @@ describe('#POST /academy/unassignTeacherFromClass', () => {
         expect(res.status).equal(200);
         expect(res.body.success).equal(true);
         expect(res.body.msg).equal('Unassign Successfully!');
-        done();
-      });
-  });
-});
-
-describe('#GET /academy/teacher/:username', () => {
-  let connect;
-  let query;
-  let username = 'quangnt';
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    query = sinon.stub(network, 'query');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    query.restore();
-  });
-
-  it('Can not connect to network', (done) => {
-    connect.returns(null);
-
-    request(app)
-      .get(`/academy/teacher/${username}`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Failed connect to blockchain!');
-        done();
-      });
-  });
-
-  it('Can not query chaincode', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    query.returns({
-      success: false,
-      msg: 'Can not query chaincode!'
-    });
-
-    request(app)
-      .get(`/academy/teacher/${username}`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Can not query chaincode!');
-        done();
-      });
-  });
-  it('do not success query teacher with student', (done) => {
-    request(app)
-      .get(`/academy/teacher/${username}`)
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-  it('success query teacher of class with admin', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify(
-      {
-        Username: 'abc',
-        Fullname: 'abcdef',
-        Info: {
-          PhoneNumber: 0376724057,
-          Email: 'abc@gmail.com',
-          Address: 'HN',
-          Sex: 'Male',
-          Birthday: 22 / 03 / 1990,
-          Avatar: 'https://images.com',
-          Country: 'VN'
-        },
-        Classes: 'ClassA'
-      },
-      {
-        Username: 'abc',
-        Fullname: 'abcdef',
-        Info: {
-          PhoneNumber: 0376724057,
-          Email: 'abc@gmail.com',
-          Address: 'HN',
-          Sex: 'Male',
-          Birthday: 22 / 03 / 1990,
-          Avatar: 'https://images.com',
-          Country: 'VN'
-        },
-        Classes: 'ClassA'
-      }
-    );
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .get(`/academy/teacher/${username}`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
         done();
       });
   });
